Make example "layereditor" to be a standalone CMake project
