+++ examples/graphicsproxy/graphicsproxycore/CMakeLists.txt
@@ -1,8 +1,10 @@
 # library: celleditorscore
 set(library_name graphicsproxycore)
 
+set(CMAKE_AUTOMOC ON)
+
 add_library(${library_name} STATIC "")
-target_link_libraries(${library_name} PRIVATE mvvm_view)
+target_link_libraries(${library_name} PRIVATE MVVM::View)
 
 target_sources(${library_name} PRIVATE
     axesrectangleview.cpp
