+++ examples/layereditor/layereditorcore/model/applicationmodels.cpp
