+++ examples/layereditor/layereditorcore/model/applicationmodels.h
