+++ examples/layereditor/layereditorcore/model/item_constants.h
