+++ examples/layereditor/layereditorcore/model/layeritems.h
