+++ examples/layereditor/layereditorcore/model/materialitems.h
