+++ examples/layereditor/layereditorcore/model/materialmodel.h
