+++ examples/layereditor/layereditorcore/model/materialpropertycontroller.h
