+++ examples/layereditor/layereditorcore/model/samplemodel.h
