+++ examples/layereditor/layereditorcore/viewmodel/CMakeLists.txt
