+++ examples/layereditor/layereditorcore/viewmodel/customeditorfactory.h
