+++ examples/layereditor/layereditorcore/viewmodel/custommodeldelegate.h
