+++ examples/layereditor/layereditorcore/viewmodel/layertableviewmodel.h
