+++ examples/layereditor/layereditorcore/viewmodel/layertableviewmodelcontroller.h
