+++ examples/layereditor/layereditorcore/viewmodel/materialtableviewmodel.h
