+++ examples/layereditor/layereditorcore/widgets/CMakeLists.txt
