+++ examples/layereditor/layereditorcore/widgets/layertablewidget.cpp
@@ -10,12 +10,12 @@
 #include <QHeaderView>
 #include <QTreeView>
 #include <QVBoxLayout>
-#include <layereditor/model/applicationmodels.h>
-#include <layereditor/model/layeritems.h>
-#include <layereditor/model/samplemodel.h>
-#include <layereditor/viewmodel/custommodeldelegate.h>
-#include <layereditor/viewmodel/layertableviewmodel.h>
-#include <layereditor/widgets/layertablewidget.h>
+#include <layereditorcore/model/applicationmodels.h>
+#include <layereditorcore/model/layeritems.h>
+#include <layereditorcore/model/samplemodel.h>
+#include <layereditorcore/viewmodel/custommodeldelegate.h>
+#include <layereditorcore/viewmodel/layertableviewmodel.h>
+#include <layereditorcore/widgets/layertablewidget.h>
 #include <mvvm/model/modelutils.h>
 #include <mvvm/model/sessionitem.h>
 #include <mvvm/viewmodel/viewmodel.h>
