+++ examples/layereditor/layereditorcore/widgets/layertablewidget.h
