+++ examples/layereditor/layereditorcore/widgets/mainwindow.h
