+++ examples/layereditor/layereditorcore/widgets/materialtablewidget.h
