+++ examples/layereditor/layereditorcore/widgets/samplewidget.h
